requirements.txt
PyExifTool
tqdm
argparse